# release-2.1: opt: Improve list and optbuilder perf