… pkg/sql/opt/optbuilder/join.go
… pkg/sql/opt/optbuilder/testdata/join